refactor dependencies between instance-related parts: add zeoclientX-base
Introduce a zeoclientX-base part that is not supposed to be the base of
zeoclient_debug. This avoids having to change common settings such as
additional eggs or environment variables for both zeoclient1 and
zeoclient2. Pulling in one part into the other as a template (<=) doesn't
help with such overridings, and extending options (+=) within a template
part may even render .installed.cfg unreadable by buildout itself.